working state

+++ src/main/java/org/aksw/faraday_cage/engine/AdjacencyMatrix.java
@@ -15,10 +15,10 @@ class AdjacencyMatrix {
 
   private static final Logger logger = LoggerFactory.getLogger(AdjacencyMatrix.class);
 
-  private final short[][][] rows;
+  private final int[][][] rows;
 
   AdjacencyMatrix(int n) {
-    rows = new short[n][n][0];
+    rows = new int[n][n][0];
   }
 
   AdjacencyMatrix(AdjacencyMatrix other) {
@@ -26,76 +26,76 @@ class AdjacencyMatrix {
   }
 
   void addEdge(int from, int fromPort, int to, int toPort) {
-    short k = (short) rows[from][to].length;
+    int k = rows[from][to].length;
     rows[from][to] = Arrays.copyOf(rows[from][to], k+2);
-    rows[from][to][k] = (short) fromPort;
-    rows[from][to][k+1] = (short) toPort;
+    rows[from][to][k] = fromPort;
+    rows[from][to][k+1] = toPort;
   }
 
   /**
    * Try to generate canonical indexing, abort if a cycle gets detected
    */
   <T> ExecutionGraph<T> compileCanonicalForm(List<ExecutionNode<T>> ops) {
-    short[] canonicalIndexing = getCanonicalIndexing();
+    int[] canonicalIndexing = getCanonicalIndexing();
     List<ExecutionNode<T>> canonicalOps = new ArrayList<>(ops.size());
     for (int i = 0; i < rows.length; i++) canonicalOps.add(null);
     AdjacencyMatrix canonicalForm = new AdjacencyMatrix(rows.length);
 
-    for (short i = 0; i < rows.length; i++) {
-      short k = (short) (rows.length-1-canonicalIndexing[i]);
+    for (int i = 0; i < rows.length; i++) {
+      int k = rows.length-1-canonicalIndexing[i];
       canonicalOps.set(k, ops.get(i));
-      for (short j = 0; j < rows.length; j++) {
-        short l = (short) (rows.length-1-canonicalIndexing[j]);
+      for (int j = 0; j < rows.length; j++) {
+        int l = rows.length-1-canonicalIndexing[j];
         canonicalForm.rows[k][l] = rows[i][j];
       }
     }
 
-    short[] inDegrees = new short[rows.length];
-    short[] outDegrees = new short[rows.length];
-    for (short i = 0; i < rows.length; i++) {
-      for (short j = 0; j < rows.length; j++) {
+    int[] inDegrees = new int[rows.length];
+    int[] outDegrees = new int[rows.length];
+    for (int i = 0; i < rows.length; i++) {
+      for (int j = 0; j < rows.length; j++) {
         inDegrees[i] += canonicalForm.rows[j][i].length/2;
         for (int k = 0; k < canonicalForm.rows[i][j].length; k+=2) {
-          outDegrees[i] = (short) Math.max(outDegrees[i], canonicalForm.rows[i][j][k]+1);
+          outDegrees[i] = Math.max(outDegrees[i], canonicalForm.rows[i][j][k]+1);
         }
       }
     }
 
     ExecutionGraph<T> executionGraph = new ExecutionGraph<>(canonicalOps.size());
-    for (short i = (short)(rows.length - 1); i >= 0; i--) {
-      short[] columnRow = new short[2+inDegrees[i]*2];
+    for (int i = rows.length - 1; i >= 0; i--) {
+      int[] columnRow = new int[2+inDegrees[i]*2];
       columnRow[0] = inDegrees[i];
       columnRow[1] = outDegrees[i];
-      for (short j = 0; j < rows.length; j++) {
-        short n = (short) canonicalForm.rows[j][i].length;
-        for (short k = 0; k < n; k+=2) {
-          columnRow[2+canonicalForm.rows[j][i][k+1]*2] = (short)(rows.length - j -1);
+      for (int j = 0; j < rows.length; j++) {
+        int n = canonicalForm.rows[j][i].length;
+        for (int k = 0; k < n; k+=2) {
+          columnRow[2+canonicalForm.rows[j][i][k+1]*2] = rows.length - j -1;
           columnRow[2+canonicalForm.rows[j][i][k+1]*2+1] = canonicalForm.rows[j][i][k];
         }
       }
-      executionGraph.addRow((short)(rows.length - i -1), canonicalOps.get(i), columnRow);
+      executionGraph.addRow(rows.length - i -1, canonicalOps.get(i), columnRow);
     }
     logger.info("\n{}", executionGraph);
     return executionGraph;
   }
 
-  private short[] getCanonicalIndexing() {
+  private int[] getCanonicalIndexing() {
     AdjacencyMatrix a = this;
-    short[] maxPathLength = new short[rows.length];
-    short[] canonicalIndexing = new short[rows.length];
-    List<Short> leafs = getLeafs();
+    int[] maxPathLength = new int[rows.length];
+    int[] canonicalIndexing = new int[rows.length];
+    List<Integer> leafs = getLeafs();
     logger.trace(leafs.toString());
-    for (short n = 1; n <= rows.length; n++) {
+    for (int n = 1; n <= rows.length; n++) {
       logger.trace("\n{}",a);
-      short[][][] x = a.rows;
-      for (short i = 0; i < rows.length; i++) {
+      int[][][] x = a.rows;
+      for (int i = 0; i < rows.length; i++) {
         if (x[i][i].length != 0) {
           throw new InvalidExecutionGraphException("The given graph contains a cycle!");
         }
       }
       boolean stop = true;
-      for (short leaf : leafs) {
-        for (short j = 0; j < rows.length; j++) {
+      for (int leaf : leafs) {
+        for (int j = 0; j < rows.length; j++) {
           if (x[leaf][j].length != 0) {
             maxPathLength[j] = n;
             stop = false;
@@ -107,9 +107,9 @@ class AdjacencyMatrix {
       }
       a = this.matrixProduct(a);
     }
-    short k = 0; short l = 0;
+    int k = 0; int l = 0;
     while (k < rows.length && l < rows.length) {
-      for (short i = 0; i < rows.length; i++) {
+      for (int i = 0; i < rows.length; i++) {
         if (maxPathLength[i] == k) {
           canonicalIndexing[i] = l;
           l++;
@@ -124,14 +124,14 @@ class AdjacencyMatrix {
 
   private AdjacencyMatrix matrixProduct(AdjacencyMatrix other) {
     AdjacencyMatrix result = new AdjacencyMatrix(rows.length);
-    short[][][] a = rows;
-    short[][][] b = other.rows;
-    short[][][] c = result.rows;
-    for (short i = 0; i < c.length; i++) {
-      for (short j = 0; j < c.length; j++) {
-        for (short k = 0; k < c.length; k++) {
+    int[][][] a = rows;
+    int[][][] b = other.rows;
+    int[][][] c = result.rows;
+    for (int i = 0; i < c.length; i++) {
+      for (int j = 0; j < c.length; j++) {
+        for (int k = 0; k < c.length; k++) {
           if ((a[i][k].length & b[k][j].length) != 0) {
-            c[i][j] = new short[2];
+            c[i][j] = new int[2];
             break;
           }
         }
@@ -140,11 +140,11 @@ class AdjacencyMatrix {
     return result;
   }
 
-  private List<Short> getLeafs() {
-    List<Short> leafs = new ArrayList<>();
-    for (short i = 0; i < rows.length; i++) {
+  private List<Integer> getLeafs() {
+    List<Integer> leafs = new ArrayList<>();
+    for (int i = 0; i < rows.length; i++) {
       boolean leaf = true;
-      for (short j = 0; j < rows.length; j++) {
+      for (int j = 0; j < rows.length; j++) {
         if (rows[j][i].length != 0) {
           leaf = false;
           break;
@@ -159,8 +159,8 @@ class AdjacencyMatrix {
 
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    for (short i = 0; i < rows.length; i++) {
-      for (short j = 0; j < rows.length; j++) {
+    for (int i = 0; i < rows.length; i++) {
+      for (int j = 0; j < rows.length; j++) {
         sb.append(rows[i][j].length/2);
         sb.append(" ");
       }
